requirements.txt
Django==2.2
django-filter==2.0.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
djangorestframework-stubs==0.2.0
ipython==7.1.1
ipython-genutils==0.2.0
itypes==1.1.0
Jinja2==2.10
PyJWT==1.7.1
PyMySQL==0.8.0
pytz==2018.9
PyYAML==3.13
redis==2.10.6
requests==2.21.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
six==1.12.0
django-oauth-toolkit==1.1.0